+++ blender-qmm/mats/extras/RedMetal.py
@@ -1,5 +1,5 @@
 import bpy
-import time 
+import time
 
 # MESSAGE BOX
 message_text = "This material already exists"
@@ -10,9 +10,8 @@ def ShowMessageBox(message="", title="", icon='INFO'):
         self.layout.label(text=message)
     bpy.context.window_manager.popup_menu(draw, title=title, icon=icon)
 
+
 # RedMetalShaderOperator
-
-
 class QMMRedMetal(bpy.types.Operator):
     """Add/Apply Red Metal Material to Selected Object (or Scene)"""
     bl_label  = "QMM Red Metal Shader"
@@ -22,12 +21,21 @@ class QMMRedMetal(bpy.types.Operator):
         # DOES THE MATERIAL ALREADY EXIST?
         if m_red_metal := bpy.data.materials.get("QMM Red Metal"):
             ShowMessageBox(message_text, "QMM Red Metal")
-            # print(f"QMM Red Metal already exists")
             bpy.context.object.active_material = m_red_metal
             return {'FINISHED'}
         else:
             self.make_shader()
         return {'FINISHED'}
+
+    def make_colorramp(self, nodes, arg1, locY, ramp_colors):
+        result = nodes.new("ShaderNodeValToRGB")
+        result.label = arg1
+        result.location = (-800, locY)
+        result.color_ramp.elements[0].color = ramp_colors[0]
+        result.color_ramp.elements[1].color = ramp_colors[1]
+        result.color_ramp.elements[0].position = 0.2
+        result.color_ramp.elements[1].position = 0.6
+        return result
 
     def make_shader(self):
         start = time.time()
@@ -54,29 +62,13 @@ class QMMRedMetal(bpy.types.Operator):
         BSDF.inputs[9].default_value = 0.4
         BSDF.inputs[16].default_value = 3.256
 
-        # colorramp
-        m_colorramp = nodes.new('ShaderNodeValToRGB')
-        m_colorramp.label = "Cinnabar"
-        m_colorramp.location = (-800, -200)
-        m_colorramp.color_ramp.elements[0].color = (
-            0.768151, 0.054480, 0.034340, 1)
-        m_colorramp.color_ramp.elements[1].color = (
-            0.964687, 0.066626, 0.002428, 1)
-        m_colorramp.color_ramp.elements[0].position = 0.2
-        m_colorramp.color_ramp.elements[1].position = 0.6
+        # colorramp cinnabar
+        c_cinnabar = [(0.768151, 0.054480, 0.034340, 1), (0.964687, 0.066626, 0.002428, 1)]
+        m_colorramp = self.make_colorramp(nodes, "Cinnabar", -200, c_cinnabar)
 
-        # colorrampvermillion
-        m_colorramp2 = nodes.new('ShaderNodeValToRGB')
-        m_colorramp2.label = "Vermillion"
-        m_colorramp2.location = (-800, 100)
-        m_colorramp2.color_ramp.elements[0].color = (
-            0.964687, 0.066626, 0.002428, 1)
-        m_colorramp2.color_ramp.elements[1].color = (
-            0.473532, 0.152926, 0.001518, 1)
-        m_colorramp2.color_ramp.elements[0].position = 0.2
-        m_colorramp2.color_ramp.elements[1].position = 0.6
-
-        nodes = m_red_metal.node_tree.nodes
+        # colorramp vermillion
+        c_vermillion = [(0.964687, 0.066626, 0.002428, 1), (0.473532, 0.152926, 0.001518, 1)]
+        m_colorramp2 = self.make_colorramp(nodes, "Vermillion", 100, c_vermillion)
 
         # EnergyConservationGroup
         bpy.ops.node.ec_group_operator()
@@ -94,8 +86,7 @@ class QMMRedMetal(bpy.types.Operator):
         texturizer_group.node_tree = bpy.data.node_groups['Texturizer']
         texturizer_group.location = (-1100, -400)
         texturizer_group.width = 240
-        texturizer_group.inputs[0].default_value = (
-            0.768151, 0.054480, 0.034340, 1)
+        texturizer_group.inputs[0].default_value = (0.768151, 0.054480, 0.034340, 1)
         texturizer_group.inputs[1].default_value = 0.4
         texturizer_group.inputs[4].default_value = 0.0125